// src/types/character.ts
export interface Character {
	id: number
	name: string
	ki: string
	maxKi: string
	race: string
	gender: 'Male' | 'Female' | 'Other'
	description: string
	image: string
	affiliation: string
	deletedAt: string | null
}

export interface CharactersResponse {
	items: Character[]
	meta: {
		totalItems: number
		itemCount: number
		itemsPerPage: number
		totalPages: number
		currentPage: number
	}
	links: {
		first: string
		previous: string
		next: string
		last: string
	}
}